using entt system

+++ Unnamed/src/EnemyObject.cpp
@@ -11,7 +11,7 @@ void EnemyObject::Init(sf::Texture& texture, WayPoint* wps, const sf::Vector2f s
 {
 	_sprite.setTexture(texture);
 	_sprite.setScale(1.f, 1.f);
-	_sprite.setOrigin(texture.getSize().x / 2, texture.getSize().y / 2);
+	_sprite.setOrigin(float(texture.getSize().x / 2), float(texture.getSize().y / 2));
 	_sprite.setPosition(spawnPos);
 	_state.live.stats.DEFAULT_HP = 100.f;
 	_state.live.stats.DEFAULT_SPD = 500.f;
